surebet_bot/requirements.txt

# Dépendances Bot Surebet VDO Group

aiohttp>=3.9.0
aiosqlite>=0.19.0
streamlit>=1.30.0
pandas>=2.0.0
plotly>=5.18.0
requests>=2.31.0
python-dotenv>=1.0.0
scrapling
faker
